Optimal policy scenario, Rt, altruism and bug fix
The code is adapted to include the optimal scenario, to calculate and plot the effective reproduction number over time and to include a parameter alpha, which can vary from completely selfish individuals (laissez-faire), to completely altruistic individuals (optimal-policy). Apart from this, it was fixed that for very low levels (<1) of Ni, a_t would stay at a minimum level to prevent the disease to spread again. When this happens now, transitions is put equal to n, which results in the activity being reset to 1 and the disease to have been completely eradicated